move : SnorkelingPointList 컴포넌트 이동
