+++ __tests__/components/unit/snorkeling-point/snorkelinPointList.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import Slider from 'react-slick';
 import { MemoryRouterProvider } from 'next-router-mock/MemoryRouterProvider';
-import SnorkelingPointListUI from '@components/unit/snorkeling-point/snorkelingPointList.presenter';
+import SnorkelingPointListUI from '@components/unit/snorkeling-point/list/snorkelingPointList.presenter';
 
 jest.mock('react-slick');
 describe('LoginContainer', () => {
