+++ pages/snorkeling-points/index.tsx
@@ -1,4 +1,4 @@
-import SnorkelingPointListUI from '@components/unit/snorkeling-point/snorkelingPointList.presenter';
+import SnorkelingPointListUI from '@components/unit/snorkeling-point/list/snorkelingPointList.presenter';
 
 export default function SnorkelingPointsPage() {
   return <SnorkelingPointListUI />;
